Finishing touches on the new method of setting tasks based off project scope
