+++ app/src/main/java/com/c0llabor8/kanban/fragment/TaskListFragment.java
@@ -1,28 +1,21 @@
 package com.c0llabor8.kanban.fragment;
 
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.databinding.DataBindingUtil;
-import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import com.c0llabor8.kanban.R;
 import com.c0llabor8.kanban.adapter.TaskListAdapter;
 import com.c0llabor8.kanban.databinding.FragmentTaskListBinding;
 import com.c0llabor8.kanban.fragment.base.BaseTaskFragment;
 import com.c0llabor8.kanban.model.Task;
-import com.parse.FindCallback;
-import com.parse.ParseException;
-import java.util.ArrayList;
-import java.util.List;
 
-public class TaskListFragment extends Fragment {
+public class TaskListFragment extends BaseTaskFragment {
 
-  private List<Task> taskList;
   private TaskListAdapter listAdapter;
   private FragmentTaskListBinding binding;
 
@@ -36,28 +29,28 @@ public class TaskListFragment extends Fragment {
     return fragment;
   }
 
-  @Override
-  public void onCreate(@Nullable Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    taskList = getArguments().getParcelableArrayList("tasks");
-  }
-
   @Nullable
   @Override
   public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
       @Nullable Bundle savedInstanceState) {
 
-    if (taskList == null) {
+    if (isPersonal) {
       Task.queryUserTasks((objects, e) -> {
-        taskList = objects;
-
-        listAdapter = new TaskListAdapter(taskList);
-        binding.rvTasks.setLayoutManager(new LinearLayoutManager(getContext()));
-        binding.rvTasks.setAdapter(listAdapter);
+        taskList.addAll(objects);
+        listAdapter.notifyDataSetChanged();
       });
     }
 
     binding = DataBindingUtil.inflate(inflater, R.layout.fragment_task_list, container, false);
     return binding.getRoot();
+  }
+
+  @Override
+  public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+    super.onViewCreated(view, savedInstanceState);
+
+    listAdapter = new TaskListAdapter(taskList);
+    binding.rvTasks.setLayoutManager(new LinearLayoutManager(getContext()));
+    binding.rvTasks.setAdapter(listAdapter);
   }
 }
